added errors reponses

+++ controller/income.go
@@ -12,6 +12,9 @@ import (
 	"github.com/gorilla/mux"
 )
 
+var badResquest string
+var conflict string
+
 func NewIncome(w http.ResponseWriter, r *http.Request) {
 	// r.ParseForm()
 	// var err error
@@ -26,13 +29,16 @@ func NewIncome(w http.ResponseWriter, r *http.Request) {
 	// database.DB.Create(&newIncome)
 	// json.NewEncoder(w).Encode(newIncome)
 
+	badResquest = "Error 400 - Bad Request"
+	conflict = "Error 409 - Conflict"
+
 	var newIncome models.Income
 	var income []models.Income
 	json.NewDecoder(r.Body).Decode(&newIncome)
 
 	if newIncome.Id != 0 {
 		log.Println("Error 400: Bad Request - Id don't be set")
-		w.WriteHeader(400)
+		http.Error(w, badResquest+" - Id don't be set", http.StatusBadRequest)
 		return
 	}
 
@@ -42,7 +48,7 @@ func NewIncome(w http.ResponseWriter, r *http.Request) {
 	year, _ := strconv.Atoi(dates[0])
 	if (day < 1 || day > 31) || (month < 1 || month > 12) {
 		log.Println("Error 400: BAD RESQUEST - Wrong data type")
-		w.WriteHeader(400)
+		http.Error(w, badResquest+" - Wrong data type", http.StatusBadRequest)
 		return
 	}
 	newIncome.Year = year
@@ -53,7 +59,7 @@ func NewIncome(w http.ResponseWriter, r *http.Request) {
 	for i := 0; i < len(income); i++ {
 		if income[i].Describe == newIncome.Describe {
 			log.Println("Error 409: Conflict - This desccribe already exist in this month")
-			w.WriteHeader(409)
+			http.Error(w, conflict+" - This desccribe already exist for this month", http.StatusConflict)
 			return
 		}
 	}
@@ -87,6 +93,9 @@ func IncomeDetail(w http.ResponseWriter, r *http.Request) {
 	id := vars["id"]
 	var income models.Income
 	database.DB.First(&income, id)
+	if income.Id == 0 {
+		http.Error(w, "Record not found ", http.StatusNotFound)
+	}
 	json.NewEncoder(w).Encode(income)
 }
 
@@ -96,14 +105,19 @@ func EditIncome(w http.ResponseWriter, r *http.Request) {
 	var income models.Income
 	var incomeDB []models.Income
 	database.DB.First(&income, id)
+	if income.Id == 0 {
+		http.Error(w, "Record not found ", http.StatusNotFound)
+		return
+	}
 	json.NewDecoder(r.Body).Decode(&income)
+	// Tratar data na requisiçao
 	dates := strings.Split(income.Date, "-")
 	day, _ := strconv.Atoi(dates[2])
 	month, _ := strconv.Atoi(dates[1])
 	year, _ := strconv.Atoi(dates[0])
 	if (day < 1 || day > 31) || (month < 1 || month > 12) {
 		log.Println("Error 400: BAD RESQUEST - Wrong data type")
-		w.WriteHeader(400)
+		http.Error(w, badResquest+" - Wrong data type", http.StatusBadRequest)
 		return
 	}
 	income.Year = year
@@ -114,7 +128,7 @@ func EditIncome(w http.ResponseWriter, r *http.Request) {
 	for i := 0; i < len(incomeDB); i++ {
 		if incomeDB[i].Describe == income.Describe {
 			log.Println("Error 409: Conflict - This desccribe already exist in this month")
-			w.WriteHeader(409)
+			http.Error(w, conflict+" - This desccribe already exist for this month", http.StatusConflict)
 			return
 		}
 	}
@@ -127,6 +141,10 @@ func DeleteIncome(w http.ResponseWriter, r *http.Request) {
 	vars := mux.Vars(r)
 	id := vars["id"]
 	var income models.Income
+	if income.Id == 0 {
+		http.Error(w, "Record not found ", http.StatusNotFound)
+		return
+	}
 	database.DB.Delete(&income, id)
 	json.NewEncoder(w).Encode(income)
 }
